backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
websockets==12.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
ta==0.10.2
requests==2.31.0
aiofiles==23.2.1
aioredis==2.0.1
celery==5.3.4
redis==5.0.1
python-telegram-bot==20.7
MetaTrader5==5.0.45
schedule==1.2.0
pytz==2023.3
cryptography==41.0.8
bcrypt==4.1.2
email-validator==2.1.0
Jinja2==3.1.2
python-dateutil==2.8.2
asyncio-mqtt==0.16.1
websocket-client==1.6.4
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
logging==0.4.9.6